Removed old implementation.
